+++ matrix.sh
@@ -42,11 +42,11 @@ mkdir ${TMP}
 
 # collect source file
 mkidr -p ${TMP}/opt/biz
-cp  -rvf .* ${TMP}/opt/biz
+cp  -rvf ${WORKDIR}/.* ${TMP}/opt/biz
 
 # collect dependency libs
 mkdir -p ${TMP}/usr/lib/python2.7/
-cp -rvf /usr/lib/python2.7/site-packages ${TMP}/usr/lib/python2.7/
+cp -rvf ${DEP}/usr/lib/python2.7/site-packages ${TMP}/usr/lib/python2.7/
 
 # package resource
 tar zcf images.tgz -C ${TMP}  .